first

+++ start.sh
@@ -1,14 +1,18 @@
 #!/bin/sh
-# 如果文件/conf/config.php不存在，复制一波
+# 如果两个配置文件不存在，从备份中复制一波到数据卷中
 if [ ! -f /conf/config.php ]; then
 	cp /confbak/config.php /conf/config.php
 fi
 if [ ! -f /conf/.env ]; then
 	cp /confbak/.env /conf/.env
 fi
-# 下面是软连接，源-目标
-ln -s /conf/config.php /freenom/config.php
-ln -s /conf/.env /freenom/.env
+# 下面是软连接，源-目标，如果实际没有那么创建软连接
+if [ ! -f /freenom/config.php ]; then
+	ln -s /conf/config.php /freenom/config.php
+fi
+if [ ! -f /freenom/.env ]; then
+	ln -s /conf/.env /freenom/.env
+fi
 cd /freenom/
 php run
 crond
